floating point bug fixed

+++ src/main/java/com/example/geektrust/constants/FareList.java
@@ -6,7 +6,7 @@ package com.example.geektrust.constants;
          A service tax of 20% is added to the final amount.
          */
 public class FareList {
-     public static final int BASE_FARE=50;
+     public static final double BASE_FARE=50;
      public static final double ADDITIONAL_FARE_PER_KM=6.5;
      public static final double ADDITIONAL_FARE_PER_MIN=2;
      public static final double SERVICE_TAX_MULTIPLIER=0.2;
